requirements.txt
django
django-reversion
django-admin-bootstrapped
